+++ src/tests/json_examples/achievement_credential.json
@@ -1,0 +1,28 @@
+{
+    "@context": [
+      "https://www.w3.org/2018/credentials/v1",
+      "https://purl.imsglobal.org/spec/ob/v3p0/context-3.0.2.json"
+    ],
+    "id": "http://example.com/credentials/3527",
+    "type": ["VerifiableCredential", "OpenBadgeCredential"],
+    "issuer": {
+      "id": "https://example.com/issuers/876543",
+      "type": "Profile",
+      "name": "Example Corp"
+    },
+    "issuanceDate": "2010-01-01T00:00:00Z",
+    "name": "Teamwork Badge",
+    "credentialSubject": {
+      "id": "did:example:ebfeb1f712ebc6f1c276e12ec21",
+      "type": "AchievementSubject",
+      "achievement": {
+                "id": "https://example.com/achievements/21st-century-skills/teamwork",
+                "type": "Achievement",
+                "criteria": {
+                    "narrative": "Team members are nominated for this badge by their peers and recognized upon review by Example Corp management."
+                },
+                "description": "This badge recognizes the development of the capacity to collaborate within a group environment.",
+                "name": "Teamwork"
+            }
+    }
+}
